Bill S4673 version original - George Borrello
BILL NUMBER: S4673 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the election law, in relation to requirements for affi- davit ballots and absentee ballots PURPOSE: To ensure that those voters casting affidavit ballots are doing so from their proper election district. SUMMARY OF PROVISIONS: Section 1: Paragraph (c) of subdivision 2-a of section 8-302 of the election law, as amended by section 8 of part xx of chapter 55 of the laws of 2019 is amended.