Bill S4687 version original - Kevin S. Parker
BILL NUMBER: S4687 SPONSOR: PARKER TITLE OF BILL: An act to amend the penal law, in relation to making conviction of assault of a social services worker a felony PURPOSE OF BILL: Provides that conviction of assault on a social worker which prevents him or her from performing a lawful duty is a felony. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 120.08 of the penal law, as amended by chapter 476 of the laws of 2018, is amended.