Bill S4623 version original - James Sanders Jr.
BILL NUMBER: S4623 SPONSOR: SANDERS TITLE OF BILL: An act to amend the state finance law, in relation to enacting the COVID-19 essential workers appreciation act PURPOSE: To create a fund for essential workers, especially those working in and around retail outlets, to monetarily reward them for their efforts. during the Coronavirus crisis. Eligible family members of any employee of an essential business who died as a result of the contracting COVID19 will also be qualified to receive funding.