Bill S4649 version original - Timothy M. Kennedy
BILL NUMBER: S4649 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to providing for taxpayer gifts for diabetes research and education and establishing the diabetes research and education fund PURPOSE: This legislation provides for corporate franchise tax and personal income tax check-offs for taxpayer gifts for diabetes research and education. This bill also establishes the diabetes fund.