Bill S4650 version original - Timothy M. Kennedy
BILL NUMBER: S4650 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to regulating the use of unmanned aircraft in the state PURPOSE: To provide consistent regulation of unmanned aerial systems in New York consistent with Federal. Aviation Administration ("FAA") regulations and guidelines. SUMMARY OF PROVISIONS: Section 1: Adds new Article 48-D to the Vehicle and Traffic Law with sections 2500 - 2504 of the new article.