Bill S4608 version original - Andrew J. Lanza
BILL NUMBER: S4608 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to doubling the fines for traffic infractions and traffic misdemeanors committed in school traffic control zones PURPOSE: Provides that the fines for traffic infractions and traffic misdemeanors committed in a school traffic control zone are doubled. SUMMARY OF SPECIFIC PROVISIONS: Section one adds sections 1800-a and 1801-a to the vehicle and traffic law. Section two is the effective date.