Bill S4690 version original - Kevin S. Parker
BILL NUMBER: S4690 SPONSOR: PARKER TITLE OF BILL: An act to amend the public health law, in relation to authorizing the commissioner of health to make grants to be used for the costs of conducting a local blood donation drive PURPOSE OR GENERAL IDEA OF THE BILL: Provides grants to encourage community blood drives. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 3100 of the public health law is amended. Section 2: This act shall take effect immediately.