Bill S4705 version original - James Sanders Jr.
BILL NUMBER: S4705 SPONSOR: SANDERS TITLE OF BILL: An act to amend the tax law, in relation to establishing a renters' and small homeowners' tax credit PURPOSE: To provide a state tax credit to renters and small homeowners in the city of New York. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 606 of the tax law by adding a new subsection (b-1) that will provide a renters' and small homeowners' tax credit in a city of a million or more.