Bill S4632 version original - Andrew J. Lanza
BILL NUMBER: S4632 SPONSOR: LANZA TITLE OF BILL: An act to amend the public authorities law, in relation to the provision of express bus service to the south shore of Staten Island PURPOSE OR GENERAL IDEA OF BILL: To provide continued express bus Service to the downtown and midtown, business districts of Manhattan. SUMMARY OF SPECIFIC PROVISIONS: Section 1.