Bill S4636 version original - Rachel May
BILL NUMBER: S4636 SPONSOR: MAY TITLE OF BILL: An act to amend the general municipal law, in relation to including lakes in adopt-a-municipal park, shoreline or roadway programs PURPOSE: Add lakes to those areas that are eligible to participate in the Munici- pal Adoption Program. SUMMARY OF PROVISIONS: Section 277 of the General Municipal Law that provides for agreements with volunteer groups for adopt-a-municipal park, shoreline or roadway shall be amended to include lakes.