Bill S4699 version original - Anthony H. Palumbo
BILL NUMBER: S4699 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to aggravated harassment of an employee by an incarcerated individual PURPOSE: To amend the penal law in relation to aggravated harassment of an employee by an inmate to include acts or actions by an inmate which are not currently covered. SUMMARY OF PROVISIONS: Section 1 amends the first undesignated paragraph of section 240.