Bill S4689 version original - Kevin S. Parker
BILL NUMBER: S4689 SPONSOR: PARKER TITLE OF BILL: An act to amend the county law, the real property actions and proceedings law, the vehicle and traffic law, the state finance law and the judiciary law, in relation to enacting the "NY Civil Gideon Act" PURPOSE: To establish a right to counsel in civil matters involving the basic necessities of life. SUMMARY OF PROVISIONS: Section 1: This act shall be known and may be cited as the "New York Civil Gideon act".