Bill S4768 version original - Andrew J. Lanza
BILL NUMBER: S4768 SPONSOR: LANZA TITLE OF BILL: An act to amend the county law and the executive law, in relation to creating an appeals process for the district attorney PURPOSE: Relates to creating an appeals process for the district attorney. SUMMARY OF PROVISIONS: Section 1 amends Section 700 of the county law by adding a new subdivi- sion 14, which enables victims and family members of a victim to appeal the release of a person by the parole board to the District Attorney.