Bill S4755 version original - Andrew J. Lanza
BILL NUMBER: S4755 SPONSOR: LANZA TITLE OF BILL: An act to amend the tax law, in relation to enacting the "disabled child care act" PURPOSE OR GENERAL IDEA OF BILL: Provides for a tax credit for certain amount of money expended on the care of a qualifying disabled child. SUMMARY OF SPECIFIC PROVISIONS: Section 1 - This act shall be known and may be cited as the "Disabled Child Care Act". Section 2 - Section 606 of the tax law is amended by adding a new subsection (ccc).