Bill S4740A version original - James Skoufis
BILL NUMBER: S4740 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to establishing the offense of unsolicited disclosure of an intimate image PURPOSE: This legislation will make it unlawful to send unsolicited intimate images punishable by fine and allow for judicial discretion to require sexual harassment prevention classes. SUMMARY OF PROVISIONS: Section 1. Creates the crime of unsolicited disclosure of an intimate image as a violation. Section 2.