Bill S4725 version original - James Sanders Jr.
BILL NUMBER: S4725 SPONSOR: SANDERS TITLE OF BILL: An act to amend the public authorities law, in relation to the member- ship of the board of directors of the New York city water board PURPOSE: To amend the New York City Water Board to include the five NYC Borough Presidents. SUMMARY OF PROVISIONS: Section 1.