Bill S4738 version original - Timothy M. Kennedy
BILL NUMBER: S4738 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the general municipal law, in relation to participation in games of bingo by minors PURPOSE: This legislation seeks to allow an individual under the age of 18 to play bingo as long as they are accompanied by an adult. SUMMARY OF PROVISIONS: Section 1 amends Section 486 of the general municipal law to allow for an individual under the age of 18 to play bingo as long as they are accompanied by an adult.