Bill S4763 version original - Andrew J. Lanza
BILL NUMBER: S4763 SPONSOR: LANZA TITLE OF BILL: An act to amend the executive law, in relation to authorizing the flag to be flown at half-staff to commemorate the death of certain members of the armed forces PURPOSE OR GENERAL IDEA OF BILL: This bill directs that the flag be flown at half-staff to commemorate the death of members of the United States armed forces who were killed in action. SUMMARY OF PROVISIONS: Section 1.