Bill S4790 version original - Gustavo Rivera
BILL NUMBER: S4790 SPONSOR: RIVERA TITLE OF BILL: An act to amend the social services law, in relation to medical assist- ance coverage for medically tailored meals and medical nutrition therapy for the purpose of disease management PURPOSE: Provides coverage for medically tailored meals. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new paragraph (kk) to subdivision 2 of section 365-a of the social services law as it relates to medically tailored meals.