Bill S4773 version original - Andrew J. Lanza
BILL NUMBER: S4773 SPONSOR: LANZA TITLE OF BILL: An act to amend the general business law, in relation to warning labels on certain sports and muscle creams PURPOSE: Provides that any sport or muscle cream sold or given away must have a warning label if it contains methyl salicylate. SUMMARY OF PROVISIONS: Section 1 deems that this act shall be known as "Arielle's law.