Bill S4757 version original - Andrew J. Lanza
BILL NUMBER: S4757 SPONSOR: LANZA TITLE OF BILL: An act to amend the public health law, in relation to directing the commissioner of the department of health to promulgate rules and regu- lations promoting recovery from opioid misuse and reducing diversion of addiction medicines PURPOSE OR GENERAL IDEA OF BILL: Directs the New York State Health Commissioner to promulgate rules and regulation to reduce the diversion of addiction medications. SUMMARY OF PROVISIONS: Section 1.