Bill S4724 version original - James Sanders Jr.
BILL NUMBER: S4724 SPONSOR: SANDERS TITLE OF BILL: An act to amend the penal law, in relation to restricting the sale of certain lasers PURPOSE: To restrict the sale of lasers which could end up in the wrong hands and be used improperly. SUMMARY OF PROVISIONS: Section 1 amends the penal law by adding a new section 240.78. Subsection 1 states it shall be unlawful to distribute, sell, or offer for sale a laser as defined by paragraph (a) of subdivision three of section 240.