Bill S4766 version original - Andrew J. Lanza
BILL NUMBER: S4766 SPONSOR: LANZA TITLE OF BILL: An act to amend the tax law and the education law, in relation to enact- ing the "education affordability act" PURPOSE: Establishes the "education affordability act" and tax credit. SUMMARY OF PROVISIONS: SECTION 1: Title SECTION 2: Provides for definitions in Tax Law, sets limits for the maximum value of credit that may be received and income eligibility for recipients.