Bill S4734 version original - Simcha Felder
BILL NUMBER: S4734 SPONSOR: FELDER TITLE OF BILL: An act authorizing the commissioner of finance of the city of New York to establish a property tax amnesty program; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: Create a property tax amnesty program SUMMARY OF PROVISIONS: Section 1. Directs the NYC Commissioner of Finance to establish a 4-month Amnesty Program for property owners who have a delinquent tax bill.