Bill S4794 version original - Andrew Gounardes
BILL NUMBER: S4794 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the retirement and social security law, in relation to authorizing police/fire members of the New York city fire department pension fund to obtain credit for service as an EMT member SUMMARY OF PROVISIONS: Section one of this bill amends paragraph 2 of subdivision c of section 513 of the Retirement and Social Security Law. Section two sets the effective date.