Bill S4762 version original - Andrew J. Lanza
BILL NUMBER: S4762 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law, in relation to the unlawful possession of a patient record PURPOSE: This bill will criminalize the unlawful possession of a patient record. SUMMARY OF PROVISIONS: Section 1 adds a new section 250.70 to the Penal law, creating the crime of unlawful possession of a patient record, a class A misdemeanor.