Bill S1976B version A - Jessica Ramos
BILL NUMBER: S1976A SPONSOR: RAMOS TITLE OF BILL: An act to amend the criminal procedure law and the judiciary law, in relation to judicial diversion programs; and to repeal certain provisions of the criminal procedure law relating thereto SUMMARY OF PROVISIONS:: Section 1 amends the title of Article 216 of the criminal procedure law. Section 2 amends the opening paragraph of subdivision 1 of section 216.