Bill S4782 version original - Gustavo Rivera
BILL NUMBER: S4782 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to modifying the definition of telehealth provider PURPOSE: Modifies the definition of telehealth provider. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends paragraph (z) of subdivision 2 of section 2999-cc of the public health law as it relates to telehealth providers.