Bill S4796 version original - Alexis Weik
BILL NUMBER: S4796 SPONSOR: WEIK TITLE OF BILL: An act to amend the environmental conservation law, in relation to sche- duled dredging of navigable waters PURPOSE: To require the development of a regularly scheduled program of dredging the navigable waters in New York State SUMMARY OF PROVISIONS: Section 1 is the title. Section 2 states the legislative intent.