Bill S4828 version original - Andrew J. Lanza
BILL NUMBER: S4828 SPONSOR: LANZA TITLE OF BILL: An act to amend chapter 759 of the laws of 1973 relating to the transfer of lands to the United States for the establishment of the Gateway National Recreation Area, in relation to the time period in which lands may be conveyed PURPOSE: This legislation extends the time period during which lands may be transferred to U.S. for establishment of the Gateway National Recre- ational. Area.