Bill S4832 version original - Michelle Hinchey
BILL NUMBER: S4832 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the tax law, in relation to the empire state film production credit and the empire state film post production credit PURPOSE: This legislation would amend the. Film Tax credit to allow for goods and services and remove the cap on Upstate productions. SUMMARY OF PROVISIONS: This would allow for the upstate 10 percent additional credit to be used for goods and services as well as for below the line labor costs.