Bill S2964A version A - Brian Kavanagh
BILL NUMBER: S2964A SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the real property law, in relation to the timing of decisions for the sales of cooperative apartments PURPOSE: This bill would ensure that the process of purchasing cooperative hous- ing is fair and transparent by adding uniformity and predictability to the application process. SUMMARY OF PROVISIONS: Section one provides the legislative intent of the bill. Section two adds a new article 11 to the Real Property Law.