Bill S4836 version original - George Borrello
BILL NUMBER: S4836 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the state finance law, in relation to awarding grants to municipalities for capital improvement projects; and making an appropri- ation therefor PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to create a grant matching program for muni- cipalities. SUMMARY OF PROVISIONS: Section 1.