Bill S4804A version original - Anthony H. Palumbo
BILL NUMBER: S4804 SPONSOR: PALUMBO TITLE OF BILL: An act directing the department of environmental conservation to estab- lish a deer management pilot program in the town of Southold PURPOSE: A Deer Management Pilot Program is to be established within the town of Southold to limit deer over population and identify potential legisla- tive solutions. SUMMARY OF PROVISIONS: Section 1. Establishes a deer management pilot program in the town of Southold.