Bill S4813 version original - Cordell Cleare
BILL NUMBER: S4813 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to establishing a private right of action for milk price gouging PURPOSE: To expand the ability of those affected by price gouging of milk to file a lawsuit from the attorney to all private citizens. SUMMARY OF PROVISIONS: Section 1 sets forth that any individual damaged by a violation of section 396-rr may bring an action for recovery of damage, and sets guidelines for allowed compensation.