Bill S4812 version original - Cordell Cleare
BILL NUMBER: S4812 SPONSOR: CLEARE TITLE OF BILL: An act to amend the criminal procedure law, in relation to providing for the return of fines, restitution and reparation payments where there was an unjust conviction PURPOSE OF BILL: To ensure that those who are unjustly convicted receive the return of fines, restitution and reparation payments in a timely fashion. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Criminal Procedure Law CPL 420.