Bill S4856 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S4856 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to interactive gaming PURPOSE OR GENERAL IDEA OF BILL: To authorize online interactive gaming in New York State. SUMMARY OF PROVISIONS: Section 1 of this bill adds a new article 15 to the racing, pari-mutuel wagering and breeding law.