Bill S4857A version original - Robert Ortt
BILL NUMBER: S4857 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to allowing prosecutors to combine certain charges PURPOSE: Authorizes prosecutors to combine petit larceny charges occurring in one 18-month period. SUMMARY OF PROVISIONS: Section 1 amends subdivisions 9, 10, and 11 of section 155.30 of the penal law by adding the word "or" to bridge subdivisions 9, 10, 11, and 12. Section 1 also adds a new subdivision 12 to section 155.30 of the penal law.