Bill S4808 version original - Anthony H. Palumbo
BILL NUMBER: S4808 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to allowing a department of motor vehicles employee to employ an attorney-at-law to defend against certain civil actions PURPOSE: To allow a Department of Motor Vehicles employee to hire an attorney to defend against certain civil actions. SUMMARY OF PROVISIONS: Section 1.