Bill S4823 version original - James Sanders Jr.
BILL NUMBER: S4823 SPONSOR: SANDERS TITLE OF BILL: An act to amend the penal law, in relation to the unlawful use or opera- tion of an unmanned aircraft PURPOSE OR GENERAL IDEA OF THE BILL: This bill will regulate the reckless operation or use of unmanned aircrafts. SUMMARY OF PROVISIONS: Section one of the bill amends section 240.00 of the penal law as relates to the definition of an unmanned aircraft. Section two of the bill amends the penal law by adding a new section 240.