Bill S4904 version original - Dean Murray
BILL NUMBER: S4904 SPONSOR: MURRAY TITLE OF BILL: An act to amend the state finance law, in relation to prohibiting the use of state funds for non-residents seeking an abortion or related procedures PURPOSE: To prohibit the use of New York State resources and funding for medical procedures for nonresidents that result in the termination of a pregnan- cy. SUMMARY OF PROVISIONS: Section 1. amends Section 8 of the state finance law by adding a new subdivision 17 to read as follows: 17.