Bill S4867A version original - Mark Walczyk
BILL NUMBER: S4867 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the election law, in relation to exempting the county of Fulton from the requirement that at least one early voting site be located in the largest city in such county PURPOSE: This bill would exempt Fulton County from section 8-600, paragraph (b), subdivision 2 of the election law. SUMMARY OF PROVISIONS: Section 1: Amends paragraph (b) of subdivision 2 of section 8-600 of the election law.