Bill S4887 version original - Michael Gianaris
BILL NUMBER: S4887 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the labor law, in relation to enacting the "roadway excavation quality assurance act" PURPOSE: To require that workers be paid not less than a prevailing rate of wage. SUMMARY OF PROVISIONS: Section 1 provides the title of the act. Section 2 amends the labor law, adding a new section 224-f to require that workers on these projects be paid not less than a prevailing rate of wage.