Bill S4891 version original - Cordell Cleare
BILL NUMBER: S4891 SPONSOR: CLEARE TITLE OF BILL: An act to amend the education law, in relation to creating a gun violence prevention program PURPOSE: To provide for a gun violence prevention program which may be utilized in public schools. The program will be developed by the State Education Department, drawing upon the expertise of the Department of Health, the Division of Criminal Justice Services and the Department of Environ- mental Conservation.