Bill S4873 version original - Jake Ashby
BILL NUMBER: S4873 SPONSOR: ASHBY TITLE OF BILL: An act to amend the tax law, in relation to increasing the child and dependent care tax credit PURPOSE: To increase the child and dependent care tax credit to reflect increases in the cost of child and dependent care. SUMMARY OF PROVISIONS: Section 1 - amends Personal Income Tax Law to increase the maximum amount of child and dependent care costs that are subject to the child and dependent care tax credit commencing with tax year 2024.