Bill S4865 version original - Jeremy Cooney
BILL NUMBER: S4865 SPONSOR: COONEY TITLE OF BILL: An act to amend the public health law, in relation to premium payments for the child health insurance plan SUMMARY OF PROVISIONS: Section 1 amends subdivision 9 of section 2510 of the public health law, as added by section 2 of part DDD of 3 chapter 56 of the laws of 2022, to add that children who are less than four years old are also eligible for the Child Health Plus plan, regardless of household income. Section 2 is the effective date.