Bill S4901 version original - Robert Rolison
BILL NUMBER: S4901 SPONSOR: ROLISON TITLE OF BILL: An act to amend the village law and the public officers law, in relation to permitting village justices of the village of Maybrook to live in the towns of Montgomery or Hamptonburgh PURPOSE: Permits village justices of the village of Maybrook to live in the towns of Montgomery or Hamptonburgh. SUMMARY OF PROVISIONS: Section 1. Section 3-300 of the public officers law is amended by adding new subdivision 33.