Bill S4877A version original - Shelley Mayer
BILL NUMBER: S4877 SPONSOR: MAYER TITLE OF BILL: An act to amend the state finance law, in relation to the time frame for the implementation of contracts with not-for-profit corporations PURPOSE: To ensure a timely procurement process for nonprofits contracting with the state. SUMMARY OF PROVISIONS: Section one cf the bill amends section 179-q of the state finance law to define "reaistered contract.