Bill S4876A version original - Dan Stec
BILL NUMBER: S4876 SPONSOR: STEC TITLE OF BILL: An act to amend the public authorities law, in relation to the amount of St. Lawrence county economic development power that may be used by the New York Power Authority to generate net earnings PURPOSE: The purpose of this bill is to extend, for an indefinite period of time, the program funded by the sale of certain unallocated hydropower produced by the Power Authority of the State of New York (NYPA) at its St.