Bill S4915 version original - Luis R. Sepúlveda
BILL NUMBER: S4915 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the criminal procedure law, in relation to the "juvenile offender second chance act" PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to give first time juvenile offenders a second chance before putting them through a criminal conviction system so that they still have the opportunity to grow and be contributing members to society. SUMMARY OF PROVISIONS: Section 1. Subdivisions 4, 5, and 6 of section 720.