Bill S4918 version original - Luis R. Sepúlveda
BILL NUMBER: S4918 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the executive law, in relation to incarcerated individ- ual interviews by the state board of parole PURPOSE: To require in-person parole interviews instead of video interviews. SUMMARY OF PROVISIONS: Sections 1 and 2 amend Executive Law 259-i(2)(a)to require the board to conduct in-person parole interviews inside correctional facilities. Section 3 is the effective date.