Bill S4945 version original - Kevin S. Parker
BILL NUMBER: S4945 SPONSOR: PARKER TITLE OF BILL: An act to amend the state finance law, in relation to establishing a clean air account; and making an appropriation therefor PURPOSE: This bill will increase the Environmental Protection Fund (EPF) from $250 million to $300 million. This additional $50 million will be used to create a new Clean Air Account. This $50 million would be allocated as follows: 1.