Bill S4934 version original - Kevin S. Parker
BILL NUMBER: S4934 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to the "state workforce fuel reduction and conservation act" PURPOSE: Provides for a state workforce fuel reduction and conservation program which will develop and assist in implementing strategies to reduce fuel consumption. SUMMARY OF PROVISIONS: Section 1: Short title. This act shall be known and may be cited as the "state workforce fuel reduction and conservation act".