Bill S4978 version original - Dean Murray
BILL NUMBER: S4978 SPONSOR: MURRAY TITLE OF BILL: An act to amend the penal law, in relation to criminalizing the filming of a violent crime PURPOSE: To make it a crime to film the commission of a violent felony with the intent to promote or encourage the commission of the felony SUMMARY OF PROVISIONS: Section 1. Creates a new section 235.25 - creating the crime of record- ing a video in the commission of a violent felony.