Bill S4939 version original - Kevin S. Parker
BILL NUMBER: S4939 SPONSOR: PARKER TITLE OF BILL: An act to amend the environmental conservation law, in relation to establishing a product stewardship program for primary batteries PURPOSE: The purpose of this bill is to establish a battery recycling program for consumers. SUMMARY OF PROVISIONS: Section 1: Article 27 of the environmental conservation law is amended by adding a new title 16.