Bill S4911 version original - Luis R. Sepúlveda
BILL NUMBER: S4911 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the penal law, in relation to prohibiting the sale of a firearm, rifle, shotgun or ammunition on state property PURPOSE: Would prohibit the sale of a firearm, rifle, shotgun or ammunition on state property including all county, city and municipal property and any fairground located on state, county, city or municipal property. SUMMARY OF PROVISIONS: Section One: Adds a new Section 265.