Bill S4973A version original - Andrew Gounardes
BILL NUMBER: S4973 Revised 3/17/23 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the labor law, in relation to payment of wages PURPOSE OR GENERAL IDEA OF BILL: To bolster protections against wage theft for employees by clarifying that all bonuses and other forms of employment remuneration that are not purely discretionary count as wages. SUMMARY OF PROVISIONS: Section one of this bill names it the "Wage Payment Integrity Act.