Bill S4950 version original - Robert Jackson
BILL NUMBER: S4950 SPONSOR: JACKSON TITLE OF BILL: An act to amend the public health law, in relation to the employment of persons to function as infection preventionists in certain general hospitals PURPOSE: To require individuals employed as infection preventionists in hospitals to be certified and undergo continuing education. SUMMARY OF PROVISIONS: Section 1 of this bill creates a new Section 2831 in the Public Health Law to address infection preventionists. Subdivision 1 defines key terms.