Bill S4986 version original - Peter Oberacker
BILL NUMBER: S4986 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the executive law, in relation to the membership of the state board of parole SUMMARY OF PROVISIONS: Section 1 amends section 259-b of the Executive Law to require that one member of the State Parole Board be a current or former member of law enforcement and that one member be crime victim or crime victim's repre- sentative. Section 2 sets the effective date.