Bill S5002A version original - Pamela Helming
BILL NUMBER: S5002 SPONSOR: HELMING TITLE OF BILL: An act to amend the tax law, in relation to extending the authorization for Ontario county to impose additional rates of sales and compensating use taxes PURPOSE: Relates to extending the authorization for Ontario County to impose additional rates of sales and compensating use taxes. SUMMARY OF PROVISIONS: Section 1.