Bill S5001 version original - Michelle Hinchey
BILL NUMBER: S5001 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the local finance law, in relation to providing a period of probable usefulness for lead service line replacement programs as a capital asset PURPOSE: The purpose of this bill is to provide a thirty year period of probable usefulness for lead service line replacement programs as a capital asset. SUMMARY OF PROVISIONS: Section 1 amends paragraph a of § 11.