Bill S4989 version original - Peter Oberacker
BILL NUMBER: S4989 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the correction law, in relation to custody and visita- tion rights of certain sex offenders PURPOSE: To prevent physical or legal custody or unsupervised visitation of a child by sex offenders described in the provisions of the bill. Unless such rights can be proven to be in the best interest of the child.