Bill S4992 version original - Peter Oberacker
BILL NUMBER: S4992 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the education law, in relation to providing for security hardware and software safety technology PURPOSE: This bill would amend sections 753, 751, and 3602 of the education law to provide state education aid for school districts that expend resources to improve security and safety technology. SUMMARY OF PROVISIONS: More specifically, this bill would amend: 1.