Bill S4993C version original - Monica Martinez
BILL NUMBER: S4993 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to allowing the presence of dogs in state parks PURPOSE: The purpose of this legislation is to permit the presence of dogs in all state parks with the exception of the Adirondack park and Catskill park. SUMMARY OF PROVISIONS: Section 1. The parks, recreation and historic preservation law is amended by adding a new section 3.