Bill S4996 version original - Dean Murray
BILL NUMBER: S4996 SPONSOR: MURRAY TITLE OF BILL: An act to amend the tax law, in relation to the exemption of community colleges from the imposition of the metropolitan commuter transportation mobility tax PURPOSE: To exempt community colleges from the MTA payroll tax. SUMMARY OF PROVISIONS: Section 1. Amends paragraph 4 of subsection b of section 800 of the tax law as it relates to adding community colleges to the definition of "eligible educational institution".