Bill S5106 version original - Cordell Cleare
BILL NUMBER: S5106 SPONSOR: CLEARE TITLE OF BILL: An act to amend the public authorities law, in relation to prohibiting time limits or curfews on fares provided to students PURPOSE OR GENERAL IDEA OF BILL: To allow students to use their MTA reduced fare passes into the evening. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Adds a new section 1266-1 to the public authorities law, prohibits the establishment of time limits or curfews to the student fare program.