Bill S5054 version original - Andrew J. Lanza
BILL NUMBER: S5054 SPONSOR: LANZA TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to establishing an alternatively assessed student program PURPOSE: This bill would establish an alternatively assessed student program. SUMMARY OF PROVISIONS: Section one adds a new chapter 12-A, the Alternatively Assessed Student Program to Title 21-A of the Administrative Code of the City of New York. Section two is the effective date.