Bill S5095 version original - Cordell Cleare
BILL NUMBER: S5095 SPONSOR: CLEARE TITLE OF BILL: An act to amend the public health law, in relation to children and recovering mothers PURPOSE OR GENERAL IDEA OF BILL: To amend the Public Health Law by establishing the Children and Recover- ing Mothers (CHARM) act. SUMMARY OF PROVISIONS: Section 1 of the bill creates the Children and Recovering Mothers (CHARM) act. Section 2 amends Article 25 of the Public Health Law by adding a new title 4-A, in relation to children and recovering mothers.