Bill S5099 version original - Cordell Cleare
BILL NUMBER: S5099 SPONSOR: CLEARE TITLE OF BILL: An act to amend the penal law and the public authorities law, in relation to theft of certain transit services PURPOSE: The purpose of this legislation is decriminalize the offense of fare- beating by lowering the offense from a misdemeanor to a violation. SUMMARY OF PROVISIONS: Section 165.18 provision 1 states a person is guilty of theft of transit service when on a railroad or subway service without payment for trans- portation.