Bill S5055 version original - Kevin S. Parker
BILL NUMBER: S5055 SPONSOR: PARKER TITLE OF BILL: An act in relation to alternative finance investment bonds PURPOSE: To direct the chair of the Urban Development Corporation in conjunction with the chair of the Dormitory Authority, director of Budget, super- intendent of Financial Services, and the Commissioner of Tax and Finance to create an investment vehicle known as an Alternative Finance Invest- ment Bond. SUMMARY OF PROVISIONS: Section 1 sets forth the legislative findings.