Bill S5041 version original - Andrew J. Lanza
BILL NUMBER: S5041 SPONSOR: LANZA TITLE OF BILL: An act to amend the not-for-profit corporation law and the vehicle and traffic law, in relation to the operation of fire patrols; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: Authorizes certain organizations to be known law operating in the as "fire patrols" under the not-for-profit corporation county of Richmond. SUMMARY OF SPECIFIC PROVISIONS: Section 1.