Bill S5057A version original - Michelle Hinchey
BILL NUMBER: S5057 SPONSOR: HINCHEY TITLE OF BILL: An act to allow Benjamin Doty to be eligible to take the civil service exam for the position of Columbia County Sheriff PURPOSE: Authorizes Benjamin Doty to take the competitive exam for full-time. Sheriff and be placed on the eligible list for such position. SUMMARY OF PROVISIONS: Section 1. Authorizes Benjamin Doty to sit for the civil service exam and be placed on the eligibility list for full-time Sheriff. Section 2.