Bill S5036A version original - Rachel May
BILL NUMBER: S5036 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to the supervision of the manufacturing and repacking of certain medical gases or wholesaling of respiratory therapy agents PURPOSE: To conform the state requirements with those set at the federal level. SUMMARY OF PROVISIONS: Section 1 - Section 6808 of the education law is amended by adding a new subdivision 9. Section 2 - Establishes the effective date.