Bill S5072 version original - Cordell Cleare
BILL NUMBER: S5072 SPONSOR: CLEARE TITLE OF BILL: An act to amend the private housing finance law, in relation to disabled veteran access to home for heroes contracts PURPOSE: The purpose of bill is to help ensure that disabled veterans are able to make improvements to their homes to improve their accessibility. SUMMARY OF SPECIFIC PROVISIONS: Section 1 creates a new Article XXIX in the Private Housing Finance law, The New York Access to Home for Heroes Program.