Bill S5046 version original - Andrew J. Lanza
BILL NUMBER: S5046 SPONSOR: LANZA PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to conduct a comprehensive study and analy- sis of New York's state and local tax laws to provide a fair, efficient and uniform method for taxing communications providers and their services sold in the state of New York. SUMMARY OF PROVISIONS: Section 1: sets forth legislative findings.